requirements.txt
requests==2.28.1
sqlalchemy==2.0.44
python-dotenv==1.0.1
jsonschema==4.7.2
mysqlclient==2.1.0